+++ src/id_translation/fetching/_memory_fetcher.py
@@ -39,6 +39,7 @@ class MemoryFetcher(AbstractFetcher[SourceType, IdType]):
             | Mapping[SourceType, "pandas.DataFrame"]
             | Mapping[SourceType, Mapping[str, Sequence[Any]]]
         ),
+        *,
         return_all: bool = True,
         **kwargs: Unpack[AbstractFetcherParams[SourceType, IdType]],
     ) -> None: