+++ tests/fetching/test_sql_fetcher.py
@@ -133,7 +133,7 @@ def test_bad_override(column, connection_string):
     mapper: Mapper[str, str, str] = Mapper(overrides={column: "bad_column"})
     fetcher = SqlFetcher(connection_string, mapper=mapper)
     with pytest.raises(exceptions.UnknownPlaceholderError, match=repr(column)):
-        fetcher.fetch([IdsToFetch("humans", {-1})], ("id", "name"), ("id", "name"))
+        fetcher.fetch([IdsToFetch("humans", {-1})], ("id", "name"), required=("id", "name"))
     fetcher.close()